refactor scroll btn
