+++ src/App.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import './App.css';
 import {BrowserRouter, Route, Switch} from "react-router-dom" 
-import Home from "./components/Home/index"
-import Actu from "./components/Actu/index"
-import Detail from "./components/Detail/index"
+import Home from "./Components/Home/index"
+import Actu from "./Components/Actu/index"
+import Detail from "./Components/Detail/index"
 
 
 class App extends Component {
