--- src/Components/Button-up/index.css
@@ -1,50 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
-/****Button up****/
-
-.button-up{
-  position: fixed;
-  visibility: hidden;
-  left: 78%; 
-}
-
-.visible{
-  visibility: visible;
-}
-
-.button-img{
-  display: block;
-  margin-top: 80vh;
-  opacity: 0.6;
-}
-.button-img:hover{
-  opacity: 1;
-}
-
-@media (min-width: 576px) { 
-  .button-up{
-    left: 86%; 
-  } 
-
-}
-@media (min-width: 991.98px) {
-  .button-up{
-    left: 90%; 
-  } 
-
-}
- 
-
