+++ src/components/Actu/index.css
@@ -12,7 +12,7 @@
     .container-actu  {
         background: url("bkg-pc-actu.jpg") no-repeat center fixed;
         background-position-y: 0px;
-        height: 100vh;
+        height: 200vh;
     }
    
   }
