+++ src/components/Home/index.js
@@ -1,36 +1,18 @@
 import React, { Component } from "react"
 import Logo from './Logo.png';
 import "./index.css";
-import ListButton from "../../components/List-button/index"
-import Fleche from "./fleche-down.gif"
+import ListButton from "../../Components/List-button/index"
+import ButtonScroll from "../ButtonScroll/index"
 
 class Home extends Component {
   constructor(props) {
     super(props)
-    this.state = {
-      lastOffset: 0,
-      isVisible: true
-    }
+
   }
-  handleScroll(evt) {
-    const offset = Math.round(window.pageYOffset)
-    if (this.state.lastOffset !== offset) {
-      if (this.state.lastOffset > offset) {
-        this.setState({ isVisible: true, lastOffset: offset })
-      } else {
-        this.setState({ isVisible: false, lastOffset: offset })
-      }
-    }
-  }
-  componentDidMount() {
-    window.addEventListener('scroll', (evt) => this.handleScroll(evt))
-  }
-  componentWillUnmount() {
-    window.removeEventListener('scroll', (evt) => this.handleScroll(evt))
-  }
+
   render() {
     return (
-      <div className="container-home" onScroll={(evt) => this.handleScroll(evt)}>
+      <div className="container-home">
         <div className="header-description">
           <div className="logo-description">
             <img src={Logo} alt="feel'm logo" />
@@ -38,7 +20,7 @@ class Home extends Component {
           </div>
           <div className="div-question-gif">
             <p className="question">Comment vous sentez-vous ?</p>
-            <img src={Fleche} className={"gif-fleche" + (this.state.isVisible ? "" : "-close")} alt="fleche vers le bas" />
+           <ButtonScroll direction="down"/>
           </div>
         </div>
         <ListButton />
